Add initial API routes for common and automobile domains
